requirements.txt
gitpython>=2.1.11
python-gitlab>=1.6.0